18 May 2018 Anti-aliasing update
This update has four parts - two definitions here and three small code patches in the sampling area of this sketch which you can locate by searching this sketch for this date.  This update allows you to control the over-sampling periods.  Intent is to spread over samples out evenly for averaging over one period of the fundamental noise wavelength that you choose.