Limited to 3 DPots/leg, 0 to 1 HX711 outboards
Should you examine the sketch code you will or may notice it is in a state of transition.  Variable names and useages may be inconsistent and so forth.  That is because I am upgrading it with longer-term goals in mind while at the same time adding intermediate features that are easily and quickly coded.  

Longer-term upgrades are as follows:
- Am currently coding for use of less than 3 and more than 3 DPots/leg
- Am currently coding for using DPots alone, without LM334, in the bridges
- Am currently coding for using unbridged DPots, single leg only
- Am currently coding for faster conductance range changes when DUT conductance goes out of range
- Not coding for other ADC types yet, but will eventually first code for multiple HX711

These goals necessitate many smaller changes that make on uninformed observer wonder why this or that is so convoluted.  The answer in many cases is that I've learned not to make so many changes that the sketch can acquire failures/bugs in more than one or two places because troubleshooting would become harder by orders of magnitude.  So the many small modifications to support the longer term goals are being added incrementally without breaking the sketch along the way.  Thank you for your tolerance.