Had this working at one time BUT don't use for advancing project
The core of this sketch is not up to the standards of later versions that allow for better advancement opportunity.  For advancing the sketch for this project, use the deepest directory level of sketch: plant_resistance_primary_perception/sketch code/in progress/nightly_with_bugs_else_it_wouldn-t_be_here/adc_for_plant_tissue.ino as the starting point.  Thank you!  It will most certainly have bugs in it, but it is provided so you know where the sketch is headed