Code flow to best support project advancement
published  for the purpose of establishing code flow to best support project advancement.
This compiles but has algorithm logic bugs mostly relative to the fast auto-adjusting of the programmable potentiometers that need to be worked out.  HELP ME WITH THAT, PLEASE.  ANYONE?  Submit an Issue to alert me that you're working on some aspect of this.  Thank you!!!!!