Gerber files for PCB
This version has the third LED that is also in version 2, but also has VR1 and VR2 properly connected instead of being backwards.  The sketch can accommodate either direction if you modify it to do so, but this is more intuitive and according to Bourns specifications. This version also spreads the leads out for Q1-3 and gives them larger pads.  U1-2 leads also spread a bit.