You'll need this to order your own PCB
This is the gerber file zip that PCB fabricating houses need, plus a few bucks from you, to make you the PCB for this project.  This revision contains jumpered construction-time diagnostics and a jumpered (bypassable) enhanced sensitivity circuit in the front end that is designed to increase sensitivity if you supply an additional 5 volts as a negative supply and jumper accordingly.