Purchase & construct from this file. Download it first
This file doesn't run in your browser until you download it to your computer or at least re-locate it to someplace that will serve it to your browser in raw .html form.  Trying to display it in your browser from GitHub here just results in it appearing as a .txt file ready for editing if you'd have rights to edit.  Such is how GitHub works.

Once you download this file and open it, it can help track both your component ordering and your component placing on the board when they arrive.  
To order components, you'll want this file sorted by component purchasing source (vendors), so sort it that way by clicking on the column title that says "Source".   As you order each line of components, you have a convenient checkoff column titled "Sourced".  That really means that line of components has been entered into the vendor's order form.  It helps you ensure everything gets ordered once and only once.

IMPORTANT:  Do not open the sealed shipping package that U3 arrives in just because you're curious!  That component is hermetically sealed in there and needs to stay that way until immediately prior to soldering it to the board then putting dried desiccant with it and soldering the shield over it, keeping everything under that shield absolutely dry.  By that step in assembly, all the holes through the board within that shield footprint must get solder-sealed by you to complete the hermetic sealing.  If ever U3 humidity precautions are violated, you'll have to dry that component according to JEDEC standards.  (Let's not call this a life or death standard in this application - in a pinch, just dry it in an electric oven - a gas oven is too humid -  at about water's boiling temperature for long enough to restore its condition).

When components arrive and you are ready to build the project, you will sort this file in your browser on the column titled "ClickToSortForPlacement" .  Click to sort until the sorting style is the one that displays a bright white sort indicator in the title cell to the right side.  The column titled with the cell that says "Placed" helps you track which component is next in order to be soldered to the board.