Kapton tape shown but not in BOM...
Kapton tape will be needed, but since it is not a circuit component, it doesn't get into the BOM.  If you want to laser cut stencils from it, get 2" wide and 6" long minimum plus what you'll need for place assist holding tape (3/8" wide, 6' long).  Arrow has the narrow, my source for the wide was AliExpress which takes weeks.