Schematic of main board plus auxiliary devices you need to build
You'll need to build or buy a cable assembly with 1/8" plugs, preferably coaxial or at least shielded.  A set of test dongles to simulate the DUT are also called for during assembly diagnostics.  These things are on the second sheet but their references may not match the BOM.  The contacting material of the electrodes should be made of non-impregnated carbon fiber tow.  You'll probably only find prepreg unless you really search hard for the non-impregnated tow.  Where the carbon fiber connects into wire  as the circuit nears the board, it should only come into contact with gold-plated conductors.