Schematic v1.0
The optional HX711 ADC is now a break-off style tab due to lack of PCB space for it in the normal Arduino footprint.  In my judgement, the noise headroom will be inadequate for it to provide useful output in the intended application anyway.  No doubt other applications may be found that could make use of it so I will just leave it this way as an option.