Download this file to local machine b/c way GitHub delivers it
GitHub doesn't deliver this file like a web server would, so you must download it to your machine.  It is a parts list with supply houses where to purchase each component.  Components of the auxiliary board are not listed normally, but if you read the sources column you'll see the info all there just the same.