Main and aux (stacker expansion) in here are separate
This set of files have the main board separate from each other, but if you want to order and construct everything as only one board, use the files in the sub-directory appropriately named for the single combined board.  Due to the hermetic sealing requirements of the REF200 U4, you are encouraged to separate (CAREFULLY) the boards in that case and hermetically seal appropriate portions of them using the shields within minutes or hours after placing the REF200, especially in high humidity ambient conditions.