If you find it awkward to order separate boards...
These files combine main and auxiliary (stacking) boards into one if you find that more convenient to work with.  Once the components are soldered down (or before, if you prefer) the boards should eventually be separated (CAREFULLY) so the smaller one can be placed under the long shield (SHLD3) and allow complete hermetic sealing of the common space inside the three shields that the U4 (REF200) requires for long-term accuracy.  (Reading the REF200 technical data reveals it should be done within minutes or hours or you should remove, re-bake, and re-solder it.)